[test-statuses][m]: moved statuses tests from ava to mocha since travis kills some processes in ava - refs #8
